Refactor styles to own file; remove unneeded styles
