+++ libs/client-shared/src/lib/utils/map.ts
@@ -12,7 +12,7 @@ import { fromExtent as polygonFromExtent } from 'ol/geom/Polygon';
 import Map from 'ol/Map';
 import { fromLonLat, transform } from 'ol/proj';
 import { register } from 'ol/proj/proj4';
-import { Circle, Fill, RegularShape, Stroke, Style } from 'ol/style';
+import { Style } from 'ol/style';
 
 import proj4 from 'proj4';
 import { isoWGSLat, isoWGSLng } from '../models';
@@ -77,226 +77,6 @@ export const olCoordsFromLV95Array = (coords: LV95[]): Coordinate[] => coords.ma
 export const olCoordsFromLV95 = (lv95Coords: LV95): Coordinate => {
   const wgsCoords = lv95ToWGS(lv95Coords);
   return fromLonLat([isoWGSLng.unwrap(wgsCoords.lng), isoWGSLat.unwrap(wgsCoords.lat)]);
-};
-
-export const makeRhombusImage = (radius: number) =>
-  new RegularShape({
-    points: 4,
-    radius,
-    angle: 0,
-    fill: new Fill({ color: 'rgba(245, 158, 11, 1.0)' }), // no transparency on rhombus
-    stroke: new Stroke({ color: '#78350F', width: 2 }),
-  });
-
-const makeLineShape = () =>
-  new RegularShape({
-    fill: new Fill({ color: '#3B82F6' }),
-    stroke: new Stroke({ color: '#1E3A8A', width: 2 }),
-    radius: 10 / Math.SQRT2,
-    radius2: 10,
-    points: 4,
-    angle: 0,
-    scale: [1, 0.5],
-  });
-
-// Todo: z-index is off
-
-/**
- * This style is used for displaying a point study; which is used for both overview and filtered view.
- */
-const studyPoint = new Style({
-  zIndex: 3,
-  image: new Circle({
-    radius: 8,
-    fill: new Fill({ color: '#14AFB8' }),
-    stroke: new Stroke({ color: '#13474E', width: 2 }),
-  }),
-});
-
-/**
- * These styles are used in the non-filtered overview and they consist of point representations (location _or_ centroid)
- */
-const overviewStyles = {
-  studyOverviewPoint: studyPoint,
-  studyOverviewPolygon: new Style({
-    zIndex: 1,
-    image: makeRhombusImage(5),
-  }),
-  studyOverviewLine: new Style({
-    zIndex: 2,
-    image: makeLineShape(),
-  }),
-};
-
-/**
- * These styles are used for the filtered view; i.e. they also style the geometries and not just points.
- */
-const filteredStyles = {
-  filteredPoint: studyPoint,
-  filteredPolygon: new Style({
-    fill: new Fill({ color: 'rgba(245, 158, 11, 0.2)' }),
-    stroke: new Stroke({ color: '#78350F', width: 2 }),
-    zIndex: 1,
-  }),
-  filteredLine: [
-    // border style (slightly thicker)
-    new Style({
-      stroke: new Stroke({
-        color: '#1E3A8A',
-        width: 10,
-      }),
-      zIndex: 2,
-    }),
-    // fill style
-    new Style({
-      stroke: new Stroke({
-        color: '#3B82F6',
-        width: 6,
-      }),
-      zIndex: 2,
-    }),
-  ],
-};
-// todo: check where selectedPolygons are used
-/**
- * These styles are used for the selected asset
- */
-const selectedStyles = {
-  selectedPoint: [
-    new Style({
-      zIndex: 3,
-      image: new Circle({
-        radius: 16,
-        fill: new Fill({ color: '#EC4899' }),
-      }),
-    }),
-    studyPoint,
-  ],
-  selectedPolygon: new Style({
-    zIndex: 999,
-    stroke: new Stroke({ color: '#EC4899', width: 8 }),
-  }),
-  selectedLine: new Style({
-    zIndex: 2,
-    stroke: new Stroke({ color: 'red', width: 3 }),
-  }),
-};
-
-/**
- * These styles are used when hovering an asset
- */
-const hoveredStyles = {
-  hoveredPoint: [
-    new Style({
-      zIndex: 3,
-      image: new Circle({
-        radius: 16,
-        fill: new Fill({ color: '#8B5CF6' }),
-      }),
-    }),
-    studyPoint,
-  ],
-  hoveredPolygon: new Style({
-    zIndex: 999,
-    stroke: new Stroke({ color: '#8B5CF6', width: 8 }),
-  }),
-  hoveredLine: new Style({
-    zIndex: 2,
-    stroke: new Stroke({ color: '#0b7285', width: 4 }),
-    fill: new Fill({ color: '#eafc5288' }),
-  }),
-};
-
-/**
- * These styles are used when editing multi-geometry objects in the editor to highlight those that are NOT currently
- * being edited.
- */
-const nonEditingStyles = {
-  nonEditingPoint: [
-    new Style({
-      zIndex: 3,
-      image: new Circle({
-        radius: 16,
-        stroke: new Stroke({ color: '#ff0000', width: 6, lineDash: [10, 5] }),
-        fill: new Fill({ color: '#ffffff88' }),
-      }),
-    }),
-    studyPoint,
-  ],
-  nonEditingPolygon: [
-    new Style({
-      zIndex: 1,
-      stroke: new Stroke({ color: '#ff0000', width: 3, lineDash: [10, 10] }),
-      fill: new Fill({ color: '#ffffff88' }),
-    }),
-  ],
-  nonEditingLine: [
-    new Style({
-      zIndex: 2,
-      stroke: new Stroke({ color: '#ff0000', width: 3, lineDash: [10, 10] }),
-    }),
-  ],
-};
-
-export const featureStyles = {
-  hidden: new Style(undefined),
-  ...overviewStyles,
-  ...filteredStyles,
-  ...selectedStyles,
-  ...hoveredStyles,
-  ...nonEditingStyles,
-};
-
-/**
- * Legacy styles used within the editor
- */
-export const editorStyles = {
-  bigPointAsset: new Style({
-    image: new Circle({
-      radius: 20,
-      stroke: new Stroke({ color: 'red', width: 6 }),
-      fill: new Fill({ color: '#ffffff88' }),
-    }),
-  }),
-  bigPointAssetHighlighted: new Style({
-    image: new Circle({
-      radius: 20,
-      stroke: new Stroke({ color: '#0b7285', width: 6 }),
-      fill: new Fill({ color: '#eafc5288' }),
-    }),
-  }),
-  bigPointAssetNotSelected: new Style({
-    image: new Circle({
-      radius: 20,
-      stroke: new Stroke({ color: '#ff0000', width: 6, lineDash: [10, 5] }),
-      fill: new Fill({ color: '#ffffff88' }),
-    }),
-  }),
-  polygonAsset: new Style({
-    stroke: new Stroke({ color: 'red', width: 3 }),
-    fill: new Fill({ color: '#ffffff88' }),
-  }),
-  polygonAssetHighlighted: new Style({
-    stroke: new Stroke({ color: '#0b7285', width: 4 }),
-    fill: new Fill({ color: '#eafc5288' }),
-  }),
-  polygonAssetNotSelected: new Style({
-    stroke: new Stroke({ color: '#ff0000', width: 3, lineDash: [10, 10] }),
-    fill: new Fill({ color: '#ffffff88' }),
-  }),
-  lineStringAsset: new Style({
-    stroke: new Stroke({ color: 'red', width: 3 }),
-  }),
-  lineStringAssetHighlighted: new Style({
-    stroke: new Stroke({ color: '#0b7285', width: 4 }),
-    fill: new Fill({ color: '#eafc5288' }),
-  }),
-  lineStringAssetNotSelected: new Style({
-    stroke: new Stroke({ color: '#ff0000', width: 3, lineDash: [10, 10] }),
-  }),
-  linePreview: new Style({
-    stroke: new Stroke({ color: '#0b7285', width: 3, lineDash: [10, 10] }),
-  }),
 };
 
 export const zoomToStudies = (
